utils/signedurls/requirements.txt
uvicorn==0.18.3
requests==2.28.1
httpx==0.23.1
gunicorn
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
fastapi==0.85.0
